fix: address CodeQL alerts (workflow permissions, int-overflow-before-widen)
- test-macos.yml: add explicit minimal permissions block (contents: read)
- frozen2_value_serializer.h: widen len to size_t before multiplying by
  elem_size so the resize size computation cannot overflow uint32_t
- compression_algorithm_benchmark.cpp: widen width to size_t before
  multiplying by height in the pixel count computation

(fixes code-scanning alerts #41, #42, #43)

+++ include/dwarfs/metadata/legacy/frozen2_value_serializer.h
@@ -201,7 +201,7 @@ void Serializer::serialize_vector(
   if (elem_layout && !elem_layout->is_none() && !vec.empty()) {
     uint16_t elem_size = elem_layout->byte_size();
     size_t new_base = buf_.size();
-    buf_.resize(new_base + len * elem_size, 0);
+    buf_.resize(new_base + static_cast<size_t>(len) * elem_size, 0);
 
     Serializer elem_ser(elem_layout, buf_, static_cast<uint32_t>(new_base),
                         static_cast<uint32_t>(new_base));
